py/ai/fininsightgpt/src/requirements.txt
pymupdf>=1.22.5
python-docx>=0.8.11
python-pptx>=0.6.21
pandas>=2.0.0
openpyxl>=3.1.2
Pillow>=10.0.0
pytesseract>=0.3.10
openai>=1.3.0
tiktoken>=0.5.0
python-dotenv>=1.0.0